Add missing comma and clarify import information
The example code for creating a Training Set is not readily usable by those who might wish to copy/paste it or use it as a template. I have taken strides here to improve its usability as such.

There was also a syntax error from a missing comma, so I put that in there, too.